Implement ReminderDao
